refactor: Entity <-> DTO 전환 로직을 {domain}Mapper class로 중앙화
