+++ assignment/src/main/java/com/musinsa/showcase/application/port/dto/MinMaxProductByCategoryResponse.java
@@ -1,7 +1,6 @@
-package com.musinsa.showcase.application.port.dto.product;
+package com.musinsa.showcase.application.port.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.musinsa.showcase.application.port.dto.category.ProductOfCategoryResponse;
 
 public record MinMaxProductByCategoryResponse(
 	@JsonProperty("카테고리")
