+++ assignment/src/main/java/com/musinsa/showcase/application/port/dto/OutfitOfBrandResponse.java
@@ -1,4 +1,4 @@
-package com.musinsa.showcase.application.port.dto.brand;
+package com.musinsa.showcase.application.port.dto;
 
 import java.util.List;
