+++ assignment/src/main/java/com/musinsa/showcase/application/port/dto/OutfitOfLowestPricedBrandResponse.java
@@ -1,0 +1,10 @@
+package com.musinsa.showcase.application.port.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public record OutfitOfLowestPricedBrandResponse(
+	@JsonProperty("최저가")
+	OutfitOfBrandResponse outfitOfLowestPricedBrandResponse
+)
+{
+}
