+++ assignment/src/main/java/com/musinsa/showcase/application/port/dto/ProductOfBrandResponse.java
@@ -1,4 +1,4 @@
-package com.musinsa.showcase.application.port.dto.brand;
+package com.musinsa.showcase.application.port.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
