--- assignment/src/main/java/com/musinsa/showcase/application/port/dto/product/OutfitOfLowestPricedBrandResponse.java
@@ -1,12 +1,0 @@
-package com.musinsa.showcase.application.port.dto.product;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.musinsa.showcase.application.port.dto.brand.OutfitOfBrandResponse;
-import com.musinsa.showcase.application.port.dto.brand.ProductOfBrandResponse;
-
-public record OutfitOfLowestPricedBrandResponse(
-	@JsonProperty("최저가")
-	OutfitOfBrandResponse outfitOfLowestPricedBrandResponse
-)
-{
-}
