+++ assignment/src/main/java/com/musinsa/showcase/application/port/in/product/FindLowestAndHighestProductsByCategoryUsecase.java
@@ -1,6 +1,6 @@
 package com.musinsa.showcase.application.port.in.product;
 
-import com.musinsa.showcase.application.port.dto.product.MinMaxProductByCategoryResponse;
+import com.musinsa.showcase.application.port.dto.MinMaxProductByCategoryResponse;
 import com.musinsa.showcase.domain.Category;
 
 public interface FindLowestAndHighestProductsByCategoryUsecase {
