+++ assignment/src/main/java/com/musinsa/showcase/application/port/in/product/FindLowestPricedOutfitByBrandUsecase.java
@@ -1,6 +1,6 @@
 package com.musinsa.showcase.application.port.in.product;
 
-import com.musinsa.showcase.application.port.dto.product.OutfitOfLowestPricedBrandResponse;
+import com.musinsa.showcase.application.port.dto.OutfitOfLowestPricedBrandResponse;
 
 public interface FindLowestPricedOutfitByBrandUsecase {
 	public OutfitOfLowestPricedBrandResponse findOutfitOfLowestPricedBrand();
