+++ assignment/src/main/java/com/musinsa/showcase/application/port/in/product/FindLowestPricedOutfitByCategoryUsecase.java
@@ -1,7 +1,7 @@
 package com.musinsa.showcase.application.port.in.product;
 
-import com.musinsa.showcase.application.port.dto.product.OutfitOfLowestPricedCategoryResponse;
-import com.musinsa.showcase.application.port.dto.product.ProductResponse;
+import com.musinsa.showcase.application.port.dto.OutfitOfLowestPricedCategoryResponse;
+import com.musinsa.showcase.application.port.dto.ProductResponse;
 import com.musinsa.showcase.domain.Category;
 
 public interface FindLowestPricedOutfitByCategoryUsecase {
