+++ assignment/src/main/java/com/musinsa/showcase/application/service/CategoryService.java
@@ -8,8 +8,8 @@ import org.springframework.transaction.annotation.Transactional;
 import com.musinsa.common.exception.ApiException;
 import com.musinsa.common.exception.ErrorCode;
 import com.musinsa.common.mapper.ProductMapper;
-import com.musinsa.showcase.application.port.dto.product.OutfitOfLowestPricedCategoryResponse;
-import com.musinsa.showcase.application.port.dto.product.ProductResponse;
+import com.musinsa.showcase.application.port.dto.OutfitOfLowestPricedCategoryResponse;
+import com.musinsa.showcase.application.port.dto.ProductResponse;
 import com.musinsa.showcase.application.port.in.product.FindLowestPricedOutfitByCategoryUsecase;
 import com.musinsa.showcase.application.port.out.ReadCategoryPort;
 import com.musinsa.showcase.domain.Category;
